--- Assets/UnityDependency/Tests/SUT/Polymorphism.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 6e930e5ae8b269d49a17cab2e490f1ff
-folderAsset: yes
-DefaultImporter:
-  userData:
