--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependAbstract.cs
@@ -1,13 +1,0 @@
-﻿namespace UnityDependency.Test.SUT.Polymorphism
-{
-    using Alquimiaware;
-    using UnityEngine;
-
-    public class DependAbstract : MonoBehaviour
-    {
-        [SerializeField]
-        [Dependency(Scope.Scene)]
-        private HierarchyAbstract abstractField = null;
-        public HierarchyAbstract Abstract { get { return this.abstractField; } }
-    }
-}
