--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependAbstract.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c25dd0e841ce1c47b4fa871c4c153a9
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
