--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependAbstractDefaultAbstract.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6933236402209fd4280799f36e4b1a33
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
