--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependAbstractDefaultDerived.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8966046458c54fe4f8862a6575cd1794
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
