--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependDerived.cs
@@ -1,13 +1,0 @@
-﻿namespace UnityDependency.Test.SUT.Polymorphism
-{
-    using Alquimiaware;
-    using UnityEngine;
-
-    public class DependDerived : MonoBehaviour
-    {
-        [SerializeField]
-        [Dependency(Scope.Scene)]
-        private HierarchyDerived derived = null;
-        public HierarchyDerived Derived { get { return this.derived; } }
-    }
-}
