--- Assets/UnityDependency/Tests/SUT/Polymorphism/DependDerived.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9cde413ad70784243ae849ef16e157ef
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
