--- Assets/UnityDependency/Tests/SUT/Polymorphism/HierarchyAbstract.cs
@@ -1,18 +1,0 @@
-﻿namespace UnityDependency.Test.SUT.Polymorphism
-{
-    using Alquimiaware;
-    using UnityEngine;
-
-    public abstract class HierarchyAbstract : MonoBehaviour
-    {
-        [SerializeField]
-        [Dependency(Scope.Scene)]
-        protected BoxCollider baseProtectedField = null;
-        public BoxCollider BaseProtectedField { get { return this.baseProtectedField; } }
-
-        [SerializeField]
-        [Dependency(Scope.Scene)]
-        private CapsuleCollider basePrivateField = null;
-        public CapsuleCollider BasePrivateField { get { return this.basePrivateField; } }
-    }
-}
