--- Assets/UnityDependency/Tests/SUT/Polymorphism/HierarchyAbstract.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 294b2c12aaeae014ba24b91f7cdbafee
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
