--- Assets/UnityDependency/Tests/SUT/Polymorphism/HierarchyDerived.cs
@@ -1,13 +1,0 @@
-﻿namespace UnityDependency.Test.SUT.Polymorphism
-{
-    using Alquimiaware;
-    using UnityEngine;
-
-    public class HierarchyDerived : HierarchyAbstract
-    {
-        [SerializeField]
-        [Dependency(Scope.Scene)]
-        private SphereCollider derivedField = null;
-        public SphereCollider DerivedField { get { return this.derivedField; } }
-    }
-}
