--- Assets/UnityDependency/Tests/SUT/Polymorphism/HierarchyDerived.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2b366d4cc6ccdd941b7aad96cf402f5f
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
